# Conflicts:
#	Assets/_Le Fish/Prefabs/Conchy.prefab
#	Assets/_Le Fish/Prefabs/Food/Dough.prefab
#	Assets/_Le Fish/Prefabs/Food/Mushroom.prefab
#	Assets/_Le Fish/Prefabs/Food/Onion.prefab
#	Assets/_Le Fish/Prefabs/Food/SweetPotato.prefab
#	Assets/_Le Fish/Prefabs/Food/Tomato.prefab
#	Assets/_Le Fish/Prefabs/Food/karipap.prefab
#	Assets/_Le Fish/Prefabs/Plantgrower.prefab
#	Assets/_Le Fish/Prefabs/Rigged Plant Grower 1.prefab
#	Assets/_Le Fish/Prefabs/XR Origin.prefab

+++ Assets/_Le Fish/Scripts/ElevatorScript.cs
@@ -8,8 +8,8 @@ public class ElevatorScript : MonoBehaviour
     private Vector3 Restaurant;
     private Vector3 Menu;
     private float speed = 1f;
-    public int restaurantY;
-    public int menuY;
+    public float restaurantY;
+    public float menuY;
    [SerializeField] private bool notMoving,moveUp, moveDown;
     // Start is called before the first frame update
     void Start()
